qlog: separate out tests in purpose-focused files
