+++ qlog/src/lib.rs
@@ -661,249 +661,6 @@ impl std::fmt::Display for HexSlice<'_> {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::events::quic::PacketSent;
-    use crate::events::quic::PacketType;
-    use crate::events::quic::QuicFrame;
-    use crate::events::EventData;
-    use crate::events::RawInfo;
-    use testing::*;
-
-    #[test]
-    fn packet_sent_event_no_frames() {
-        let log_string = r#"{
-  "time": 0.0,
-  "name": "transport:packet_sent",
-  "data": {
-    "header": {
-      "packet_type": "initial",
-      "packet_number": 0,
-      "version": "1",
-      "scil": 8,
-      "dcil": 8,
-      "scid": "7e37e4dcc6682da8",
-      "dcid": "36ce104eee50101c"
-    },
-    "raw": {
-      "length": 1251,
-      "payload_length": 1224
-    }
-  }
-}"#;
-
-        let pkt_hdr = make_pkt_hdr(PacketType::Initial);
-        let ev_data = EventData::PacketSent(PacketSent {
-            header: pkt_hdr,
-            raw: Some(RawInfo {
-                length: Some(1251),
-                payload_length: Some(1224),
-                data: None,
-            }),
-            ..Default::default()
-        });
-
-        let ev = Event::with_time(0.0, ev_data);
-
-        assert_eq!(serde_json::to_string_pretty(&ev).unwrap(), log_string);
-    }
-
-    #[test]
-    fn packet_sent_event_some_frames() {
-        let log_string = r#"{
-  "time": 0.0,
-  "name": "transport:packet_sent",
-  "data": {
-    "header": {
-      "packet_type": "initial",
-      "packet_number": 0,
-      "version": "1",
-      "scil": 8,
-      "dcil": 8,
-      "scid": "7e37e4dcc6682da8",
-      "dcid": "36ce104eee50101c"
-    },
-    "raw": {
-      "length": 1251,
-      "payload_length": 1224
-    },
-    "frames": [
-      {
-        "frame_type": "padding",
-        "payload_length": 1234
-      },
-      {
-        "frame_type": "ping"
-      },
-      {
-        "frame_type": "stream",
-        "stream_id": 0,
-        "offset": 0,
-        "length": 100,
-        "fin": true
-      }
-    ]
-  }
-}"#;
-
-        let pkt_hdr = make_pkt_hdr(PacketType::Initial);
-
-        let frames = vec![
-            QuicFrame::Padding {
-                payload_length: 1234,
-                length: None,
-            },
-            QuicFrame::Ping {
-                payload_length: None,
-                length: None,
-            },
-            QuicFrame::Stream {
-                stream_id: 0,
-                offset: 0,
-                length: 100,
-                fin: Some(true),
-                raw: None,
-            },
-        ];
-
-        let ev_data = EventData::PacketSent(PacketSent {
-            header: pkt_hdr,
-            frames: Some(frames.into()),
-            raw: Some(RawInfo {
-                length: Some(1251),
-                payload_length: Some(1224),
-                data: None,
-            }),
-            ..Default::default()
-        });
-
-        let ev = Event::with_time(0.0, ev_data);
-        assert_eq!(serde_json::to_string_pretty(&ev).unwrap(), log_string);
-    }
-
-    #[test]
-    fn trace_no_events() {
-        let log_string = r#"{
-  "vantage_point": {
-    "type": "server"
-  },
-  "title": "Quiche qlog trace",
-  "description": "Quiche qlog trace description",
-  "configuration": {
-    "time_offset": 0.0
-  },
-  "events": []
-}"#;
-
-        let trace = make_trace();
-
-        let serialized = serde_json::to_string_pretty(&trace).unwrap();
-        assert_eq!(serialized, log_string);
-
-        let deserialized: Trace = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(deserialized, trace);
-    }
-
-    #[test]
-    fn trace_seq_no_events() {
-        let log_string = r#"{
-  "vantage_point": {
-    "type": "server"
-  },
-  "title": "Quiche qlog trace",
-  "description": "Quiche qlog trace description",
-  "configuration": {
-    "time_offset": 0.0
-  }
-}"#;
-
-        let trace = make_trace_seq();
-
-        let serialized = serde_json::to_string_pretty(&trace).unwrap();
-        assert_eq!(serialized, log_string);
-
-        let deserialized: TraceSeq = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(deserialized, trace);
-    }
-
-    #[test]
-    fn trace_single_transport_event() {
-        let log_string = r#"{
-  "vantage_point": {
-    "type": "server"
-  },
-  "title": "Quiche qlog trace",
-  "description": "Quiche qlog trace description",
-  "configuration": {
-    "time_offset": 0.0
-  },
-  "events": [
-    {
-      "time": 0.0,
-      "name": "transport:packet_sent",
-      "data": {
-        "header": {
-          "packet_type": "initial",
-          "packet_number": 0,
-          "version": "1",
-          "scil": 8,
-          "dcil": 8,
-          "scid": "7e37e4dcc6682da8",
-          "dcid": "36ce104eee50101c"
-        },
-        "raw": {
-          "length": 1251,
-          "payload_length": 1224
-        },
-        "frames": [
-          {
-            "frame_type": "stream",
-            "stream_id": 0,
-            "offset": 0,
-            "length": 100,
-            "fin": true
-          }
-        ]
-      }
-    }
-  ]
-}"#;
-
-        let mut trace = make_trace();
-
-        let pkt_hdr = make_pkt_hdr(PacketType::Initial);
-
-        let frames = vec![QuicFrame::Stream {
-            stream_id: 0,
-            offset: 0,
-            length: 100,
-            fin: Some(true),
-            raw: None,
-        }];
-        let event_data = EventData::PacketSent(PacketSent {
-            header: pkt_hdr,
-            frames: Some(frames.into()),
-            raw: Some(RawInfo {
-                length: Some(1251),
-                payload_length: Some(1224),
-                data: None,
-            }),
-            ..Default::default()
-        });
-
-        let ev = Event::with_time(0.0, event_data);
-
-        trace.push_event(ev);
-
-        let serialized = serde_json::to_string_pretty(&trace).unwrap();
-        assert_eq!(serialized, log_string);
-
-        let deserialized: Trace = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(deserialized, trace);
-    }
-}
-
 pub mod events;
 pub mod reader;
 pub mod streamer;
